refactor: Update task_service.go and task.go
The code changes in task_service.go and task.go refactor the way the user ID is handled. In task_service.go, the query for finding tasks now uses the user ID in hexadecimal format. In task.go, the `UserId` field is updated with the `bson` tag to omit empty values when marshaling to JSON. This improves the consistency and efficiency of the code related to user IDs.